requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
anthropic
requests
beautifulsoup4
tqdm
